core update classes

+++ fuel/core/classes/ftp.php
@@ -3,20 +3,18 @@
  * Part of the Fuel framework.
  *
  * @package    Fuel
- * @version    1.5
+ * @version    1.7
  * @author     Fuel Development Team
  * @license    MIT License
- * @copyright  2010 - 2013 Fuel Development Team
+ * @copyright  2010 - 2015 Fuel Development Team
  * @link       http://fuelphp.com
  */
 
 namespace Fuel\Core;
 
-
 class FtpConnectionException extends \FuelException {}
 
 class FtpFileAccessException extends \FuelException {}
-
 
 /**
  * FTP Class
@@ -272,7 +270,7 @@ class Ftp
 			return false;
 		}
 
-		if ( ! file_exists($local_path))
+		if ( ! is_file($local_path))
 		{
 			throw new \FtpFileAccessException('No source file');
 			return false;
@@ -459,7 +457,11 @@ class Ftp
 				// we'll recursively call delete_dir()
 				if ( ! @ftp_delete($this->_conn_id, $item))
 				{
-					$this->delete_dir($item);
+					// don't recurse into current of parent directory
+					if ( ! preg_match('/\/\.\.|\/\.$/', $item))
+					{
+						$this->delete_dir($item);
+					}
 				}
 			}
 		}
@@ -616,7 +618,7 @@ class Ftp
 			'phtml',
 			'shtml',
 			'log',
-			'xml'
+			'xml',
 		);
 
 		return in_array($ext, $text_types) ? 'ascii' : 'binary';
@@ -654,4 +656,3 @@ class Ftp
 	}
 
 }
-
